+++ src/main/java/net/jobsaddon/mixin/compat/BlockWrapperMixin.java
@@ -7,8 +7,6 @@ import net.jobsaddon.jobs.JobHelper;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.BlockPos;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -18,8 +16,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class BlockWrapperMixin {
 
     @Inject(method = "playerDestroy", at = @At("TAIL"), remap = false)
-    private void playerDestroyMixin(@NotNull ILevel level, @NotNull IPlayer player, @NotNull IBlockPos blockPos, @NotNull IBlockState blockState, @Nullable IBlockEntity blockEntity,
-                                    @NotNull IItemStack itemStack, CallbackInfo info) {
+    private void playerDestroyMixin(ILevel level, IPlayer player, IBlockPos blockPos, IBlockState blockState, IBlockEntity blockEntity, IItemStack itemStack, boolean dropResources, CallbackInfo info) {
         BlockState state = (BlockState) blockState.getRaw();
         PlayerEntity playerEntity = (PlayerEntity) player.getRaw();