+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed hardmode respawn sync
+- Fixed fallingtree compat
 ### Changed:
 -